refact(PeriodFilter): Extract common helpers to date-utils
