+++ components/host-dashboard/HostDashboardExpenses.js
@@ -6,11 +6,11 @@ import { useRouter } from 'next/router';
 import { FormattedMessage } from 'react-intl';
 
 import EXPENSE_STATUS from '../../lib/constants/expense-status';
+import { parseDateInterval } from '../../lib/date-utils';
 import { API_V2_CONTEXT, gqlV2 } from '../../lib/graphql/helpers';
 import { useLazyGraphQLPaginatedResults } from '../../lib/hooks/useLazyGraphQLPaginatedResults';
 
 import { parseAmountRange } from '../budget/filters/AmountFilter';
-import { parseDateRange } from '../budget/filters/PeriodFilter';
 import DismissibleMessage from '../DismissibleMessage';
 import ExpensesFilters from '../expenses/ExpensesFilters';
 import ExpensesList from '../expenses/ExpensesList';
@@ -124,7 +124,7 @@ const isValidStatus = status => {
 
 const getVariablesFromQuery = query => {
   const amountRange = parseAmountRange(query.amount);
-  const { from: dateFrom, to: dateTo } = parseDateRange(query.period);
+  const { from: dateFrom, to: dateTo } = parseDateInterval(query.period);
   return {
     offset: parseInt(query.offset) || 0,
     limit: (parseInt(query.limit) || NB_EXPENSES_DISPLAYED) * 2,
