+++ lib/date-utils.js
@@ -22,16 +22,6 @@ export const getNextChargeDate = (firstChargeDate, interval) => {
   }
 };
 
-/** Adds `count` months to the given `date` */
-export const addMonths = (date, count) => {
-  return new Date(date.setMonth(date.getMonth() + count));
-};
-
-/** Adds `count` years to the given `date` */
-export const addYears = (date, count) => {
-  return new Date(date.setFullYear(date.getFullYear() + count));
-};
-
 /**
  * Format a datetime object to an ISO date like `YYYY-MM-DD`
  */
@@ -86,4 +76,53 @@ export const getDateFromValue = value => {
   } else if (value instanceof dayjs) {
     return value.toDate();
   }
+};
+
+/**
+ * From a simple date as '2020-01-01', returns a string like '2020-01-01T00:00:00Z'.
+ */
+export const simpleDateToISOString = (date, isEndOfDay, timezoneType) => {
+  if (!date) {
+    return null;
+  } else {
+    const isUTC = timezoneType === 'UTC';
+    const dayjsTimeMethod = isEndOfDay ? 'endOf' : 'startOf';
+    const result = isUTC ? dayjs.utc(date) : dayjs(date);
+    return result[dayjsTimeMethod]('day').toISOString();
+  }
+};
+
+/**
+ * Parse `strValue` in a "dateFrom→dateTo" format and returns an object like { from, to, timezoneType }.
+ * Each value in the object will be `undefined` if there's no filter for it.
+ */
+export const parseDateInterval = strValue => {
+  const parsedValue = strValue?.match(/^(?<from>[^→]+)(→(?<to>.+?(?=~UTC|$)))?(~(?<timezoneType>UTC))?$/);
+  if (parsedValue) {
+    const getDateIsoString = dateStr => (!dateStr || dateStr === 'all' ? undefined : dateStr);
+    return {
+      from: getDateIsoString(parsedValue.groups.from),
+      to: getDateIsoString(parsedValue.groups.to),
+      timezoneType: parsedValue.groups.timezoneType || 'local',
+    };
+  } else {
+    return { from: undefined, to: undefined, timezoneType: 'local' };
+  }
+};
+
+/**
+ * Opposite of `parseDateInterval`: takes an object like {from, to, timezoneType} and returns a string
+ * like "from→to".
+ */
+export const encodeDateInterval = interval => {
+  if (!interval || (!interval.from && !interval.to)) {
+    return '';
+  }
+
+  const encodeDate = (date, isEndOfDay) => {
+    return simpleDateToISOString(date, isEndOfDay, interval.timezoneType) || 'all';
+  };
+
+  const baseResult = `${encodeDate(interval.from, false)}→${encodeDate(interval.to, true)}`;
+  return interval.timezoneType === 'UTC' ? `${baseResult}~UTC` : baseResult;
 };
